Added graphics to WaterContainer
Added a graphical display of the result in WaterContainer, rather than having a text-based display.  This has been done by adding a custom View to the project.
Also updated some of the original code, to allow for different sized grids, and to update some of the variable names to more common standards.