Added WithMyShadow release
Added files for a release of With My Shadow.  This has been placed in the new Nova directory.

With My Shadow consists of two zip files, one contains the installer, the other contains the files needed to run the project.  This has been done to give people a choice whether they want to install the project, or just run it from a local directory.